+++ src/larcv3/app/queueio/BatchFillerBBox.cxx
@@ -102,7 +102,7 @@ namespace larcv3 {
     */
     auto _bbox_producer      = config["Producer"].template get<std::string>();
     auto _unfilled_box_value = config["UnfilledBoxValue"].template get<float>();
-    auto _max_boxes          = config["MaxBoxes"].template         get<int>();
+    auto _max_boxes          = config["MaxBoxes"].template         get<size_t>();
     auto _slice_v            = config["Channels"].template         get<std::vector<size_t>>();
 
 
@@ -116,7 +116,7 @@ namespace larcv3 {
     //   throw larbys();
     // }
 
-    int _num_channels = dimension == 2 ? _slice_v.size() : 1;
+    size_t _num_channels = dimension == 2 ? _slice_v.size() : 1;
 
     std::vector<int> dim;
     dim.resize(4);
